fix compile warnings for gcc 6.1.1

+++ src/modules/position_estimator_inav/position_estimator_inav_main.cpp
@@ -659,7 +659,7 @@ int position_estimator_inav_thread_main(int argc, char *argv[])
 
 					double rate_threshold = 0.15f;
 
-					if (fabs(rates_setpoint.pitch) < rate_threshold) {
+                    if ((double)fabs(rates_setpoint.pitch) < rate_threshold) {
 						//warnx("[inav] test ohne comp");
 						flow_ang[0] = (flow.pixel_flow_x_integral / (float)flow.integration_timespan * 1000000.0f) * params.flow_k;//for now the flow has to be scaled (to small)
 					}
@@ -670,7 +670,7 @@ int position_estimator_inav_thread_main(int argc, char *argv[])
 							       + gyro_offset_filtered[0]) * params.flow_k;//for now the flow has to be scaled (to small)
 					}
 
-					if (fabs(rates_setpoint.roll) < rate_threshold) {
+                    if ((double)fabs(rates_setpoint.roll) < rate_threshold) {
 						flow_ang[1] = (flow.pixel_flow_y_integral / (float)flow.integration_timespan * 1000000.0f) * params.flow_k;//for now the flow has to be scaled (to small)
 					}
 					else {
@@ -681,7 +681,7 @@ int position_estimator_inav_thread_main(int argc, char *argv[])
 
 					/* flow measurements vector */
 					float flow_m[3];
-					if (fabs(rates_setpoint.yaw) < rate_threshold) {
+                    if ((double)fabs(rates_setpoint.yaw) < rate_threshold) {
 						flow_m[0] = -flow_ang[0] * flow_dist;
 						flow_m[1] = -flow_ang[1] * flow_dist;
 					} else {
